% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{auPRC_table}
\alias{auPRC_table}
\title{Return a table with model names, auPRC, PP100 and ntree of the model: only for regular rulefit, not two way}
\usage{
auPRC_table(x, ntree = x$ntree)
}
\arguments{
\item{x}{an object of class \code{varppRuleFit}}
}
\description{
Return a table with model names, auPRC, PP100 and ntree of the model: only for regular rulefit, not two way
}
