% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{auROC}
\alias{auROC}
\title{Area under the receiver operator curve plot for Class \code{varppRule}}
\usage{
auROC(x)
}
\arguments{
\item{x}{an object of class \code{varppRule}}
}
\description{
Area under the receiver operator curve plot for Class \code{varppRule}
}
