% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{density_plot}
\alias{density_plot}
\title{Density plot for the rule predictions}
\usage{
density_plot(rulefit_results)
}
\arguments{
\item{rulefit_results}{an object of class varppRuleFit}
}
\value{
a density plot for the predictions based on the final rules
}
\description{
Density plot for the rule predictions
}
