% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.class_by_threshold}
\alias{.class_by_threshold}
\title{Return model metrics for both Random Forest and RuleFit}
\usage{
.class_by_threshold(actual, predicted)
}
\arguments{
\item{actual}{values and predicted values of the model}
}
\description{
Return model metrics for both Random Forest and RuleFit
}
