% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.extract_ranger_rules}
\alias{.extract_ranger_rules}
\title{Extract rules from ranger trees}
\usage{
.extract_ranger_rules(rf_results)
}
\arguments{
\item{rf_results}{the results fro mthe ranger tree generation within the varpp function}
}
\value{
a named vector of rules
}
\description{
This function returns rules based on the decision trees built in ranger. It depends on the function varpp
}
