% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gtex.R
\name{.gtex_gene_level}
\alias{.gtex_gene_level}
\title{Read GTEx or HCL data}
\usage{
.gtex_gene_level(expression_file, benign_variant_file, pathogenic_variant_file)
}
\arguments{
\item{expression_file}{is the filepath to the expression data (GTEx or HCL)}

\item{benign_variant_file}{is the file path to the benign variant file}

\item{pathogenic_variant_file}{is the filepath to the pathogenic variant file'}
}
\value{
a list of expression data, pathogenic variants and benign variants
}
\description{
This function is needed to read in the GTEx data without two level bootstrap. Here we jsut want to see if we can
identify the disease gene
}
