% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.sample_benign_variants}
\alias{.sample_benign_variants}
\title{Sampling and subsettign for the benign variant data}
\usage{
.sample_benign_variants(benign_data, sampled_genes)
}
\arguments{
\item{benign_data}{the subset of benign variants}

\item{sampled_genes}{the sampled genes (with replacement) from the sampling step}
}
\value{
a list of randomly sampled gene variants, without replacement
}
\description{
Sampling and subsettign for the benign variant data
}
