% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{genePanelTop}
\alias{genePanelTop}
\title{Return a gene panel based on the tissues in the top rule}
\usage{
genePanelTop(varppRuleObject)
}
\arguments{
\item{varppRuleFitObject}{the results from varppRule}
}
\description{
Return a gene panel based on the tissues in the top rule
}
