% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCADDcutOff}
\alias{getCADDcutOff}
\title{Function to extract CADD score including cut-off}
\usage{
getCADDcutOff(rule)
}
\arguments{
\item{rule}{a single rule in text/string format}
}
\description{
Function to extract CADD score including cut-off
}
