% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{kappa_stats}
\alias{kappa_stats}
\title{Calculate kappa statistic}
\usage{
kappa_stats(cross_table)
}
\arguments{
\item{cross_table}{the confusion Matrix of predictions and actual data}
}
\description{
Calculate kappa statistic
}
