% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_ensemble.R
\name{lasso_ensemble}
\alias{lasso_ensemble}
\title{LASSO cross validation of rules}
\usage{
lasso_ensemble(data, rules, bootstrap.rounds, cores)
}
\arguments{
\item{data}{is a list of data with the rules added. The benign and the pathogenic variants files are necessery for the sampling.}

\item{rules}{is the list of rules that were generated in the varpp function. This is necessary for the annotation of the final results.}

\item{bootstrap.rounds}{number of bootstrap rounds for the outer loop of the LASSO cross-validation, defaults to 100}

\item{cores}{number of cores for parallel, defaults to 4}
}
\value{
A list of predictions for the CADD raw rankscore and the tissue/cell specific expression added. Further, a variable importance list for all rules and variables tested.
}
\description{
This function performs nested cross validation on the generated rule data set. It is the final step in the rulefit algorithm and returns the predictions
}
