% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{metrics}
\alias{metrics}
\title{Return model metrics for both Random Forest and RuleFit}
\usage{
metrics(x)
}
\arguments{
\item{x}{an object of class \code{varppRuleFit}}
}
\description{
Return model metrics for both Random Forest and RuleFit
}
