% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{performance_varpp}
\alias{performance_varpp}
\title{Return a table with model names, auPRC, PP100 and ntree of the model: only for two level bootstrap model}
\usage{
performance_varpp(x)
}
\arguments{
\item{x}{an object of class \code{varppRuleFit}}
}
\description{
Return a table with model names, auPRC, PP100 and ntree of the model: only for two level bootstrap model
}
