% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{predict.varppRule}
\alias{predict.varppRule}
\title{Predict function for varppRule}
\usage{
\method{predict}{varppRule}(patient_data, model_results, predict = c("probability", "class"))
}
\arguments{
\item{patient_data}{based on a patient .vcf file,
a preprocessed input file that is annotated with GTEx and CADD scores}

\item{hpo_term}{patient hpo terms}
}
\description{
This function requires extra data and can not be executed by itself. We need to download the genome_files,
prepare a patient data file from the .vcf file and can then apply this function.
}
