% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.varppRule}
\alias{print.varppRule}
\title{Class specific functions}
\usage{
\method{print}{varppRule}(x)
}
\arguments{
\item{x, }{an object of class varppRuleFit}
}
\description{
Class specific functions
}
