% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ruleVariantPlot}
\alias{ruleVariantPlot}
\title{Scatterplot of # of rules that predict correctly versus # of variants per gene}
\usage{
ruleVariantPlot(rulefit_results)
}
\arguments{
\item{rulefit_results}{an object of class varppRuleFit}

\item{y}{the outcome variable}
}
\value{
a scatterplot of # of rules and # of variants per gene
}
\description{
Scatterplot of # of rules that predict correctly versus # of variants per gene
}
