% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ruleVarImp}
\alias{ruleVarImp}
\title{Function to return all rules ranked by variable importance}
\usage{
ruleVarImp(x)
}
\arguments{
\item{x}{an object of class varppRuleFit}
}
\description{
Function to return all rules ranked by variable importance
}
