% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{selected_rule_performance}
\alias{selected_rule_performance}
\title{Prediction of single rules}
\usage{
selected_rule_performance(rulename, rulefit_results_object)
}
\arguments{
\item{rulename}{is the name of one of the rules as returned by the varppRuleFit model}

\item{rulefit_results_object, }{a varppRuleFit object}
}
\description{
Prediction of single rules
}
