% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data to test the rulefit function}
\usage{
simulate_data(n = 1000, p = 100, seed = 777)
}
\arguments{
\item{n}{the number of instances}

\item{p}{the number of variables}

\item{seed}{the seed, to create reproducible results}
}
\value{
a simulated data set with a binary outcome variable and 35 predictors that are associated with the outcome
}
\description{
Simulate data to test the rulefit function
}
