% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{varIMP}
\alias{varIMP}
\title{varIMP: Function to extract the variable importance of the expression data variables}
\usage{
varIMP(rule_model = NULL, HPOterm)
}
\arguments{
\item{rule_model}{This is the RuleFit model object}

\item{HPOterm}{Add the HPO term name for the model}
}
\description{
This function is provided on top of ruleVarImp. It Re-weights the rule kappas by the variabels selected per rule
and returns a 0 to 1 scaled importance value per tissue. The most important variabel will have a value of 1.
This is based on the variabel importance described in the RuleFit publication by XXX
}
