% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varpp_report.R
\name{varpp_report}
\alias{varpp_report}
\title{Report function for varpp results}
\usage{
varpp_report(results, report_filename)
}
\arguments{
\item{results}{the results from the rule_fit() function}

\item{report_filename}{The path, including the filename for the resulting report.}
}
\description{
This function will create a report based on the results from the rulef_fit() function.'
}
