% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{auPRC}
\alias{auPRC}
\title{Area under the precision recall curve plot for Class \code{varppRuleFit}}
\usage{
auPRC(x, model = c("rf", "rulefit"))
}
\arguments{
\item{x}{an object of class \code{varppRuleFit}}
}
\description{
Area under the precision recall curve plot for Class \code{varppRuleFit}
}
