% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{auROC}
\alias{auROC}
\title{Area under the receiver operator curve plot for Class \code{varppRuleFit}}
\usage{
auROC(x, model = c("rf", "rulefit"))
}
\arguments{
\item{x}{an object of class \code{varppRuleFit}}
}
\description{
Area under the receiver operator curve plot for Class \code{varppRuleFit}
}
