% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{predict.varppRule}
\alias{predict.varppRule}
\title{Predict function for varppRule}
\usage{
\method{predict}{varppRule}(patient_data, model_results, predict = c("probability", "class"))
}
\arguments{
\item{patient_data}{based on a patient .vcf file,
a preprocessed inpout file that is annotated with GTEx and CADD scores}

\item{hpo_term}{patient hpo terms}
}
\description{
Predict function for varppRule
}
