% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{removeCADD}
\alias{removeCADD}
\title{Function to remove the CADD score variable including '>' , '<' and '=' from the rules}
\usage{
removeCADD(varppRuleObject)
}
\arguments{
\item{varppRuleObject}{the results from varppRule}
}
\description{
Function to remove the CADD score variable including '>' , '<' and '=' from the rules
}
