% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varpp.R
\name{varpp}
\alias{varpp}
\title{varpp: VARiant Prioritisation by Phenotype}
\usage{
varpp(HPO, type = "gtex", ntree = 500, max.depth = NULL, cores = 4)
}
\arguments{
\item{HPO}{HPO term associated list of genes}

\item{type}{the prediction data; either hcl (single cell) or gtex (tissue specific)}

\item{ntree}{is the number of trees that should be built for ranger. It defaults to 1000}

\item{max.depth}{is the maximum tree depth for the ranger trees. IT defaults to 3.}

\item{cores}{number of cores for parallel, defaults to 4}
}
\description{
varpp: VARiant Prioritisation by Phenotype
}
